requirements.txt
absl-py==0.9.0
astor==0.8.1
beautifulsoup4==4.8.2
bleach==1.5.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cursor==1.3.4
cycler==0.10.0
deepspeech==0.6.0
deepspeech-tflite==0.6.1
enum34==1.1.6
Flask==1.1.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
halo==0.0.29
html5lib==0.9999999
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
log-symbols==0.0.14
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
nltk==3.4.5
numpy==1.14.4
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==7.0.0
playsound==1.2.2
pocketsphinx==0.1.15
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyparsing==2.4.7
pypiwin32==223
pyspellchecker==0.5.4
python-dateutil==2.8.1
pyttsx3==2.71
pywin32==227
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
spinners==0.0.24
tensorboard==2.1.0
tensorflow==1.4.0
tensorflow-estimator==2.1.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
textblob==0.15.3
tflearn==0.3.2
urllib3==1.25.8
webrtcvad==2.0.10
Werkzeug==0.16.1
wincertstore==0.2
wrapt==1.11.2
